[Networkx] typed `find_cycle`
Had to remove the @_dispatchable to use @overloads.
For the MultiGraph and Graph when given an orientation of
"reverse" or "ignore" it acts the same as "original".

This only works for Pyright. It fails with mypy due to the overloads.
I think mypy says "well a multigraph is also a graph, so the graph
overload is also available and has an incompatible return type".

Even making it less complex so the returned orientation is not
narrowed to the specific direction caused the same problem.
As soon as I add an overload for a more specific graph type it conflicts.